+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/data/view/column/ColumnView.java
@@ -19,11 +19,11 @@ import net.splitcells.dem.data.set.list.List;
 import net.splitcells.dem.data.set.list.ListView;
 import net.splitcells.dem.data.set.list.Lists;
 import net.splitcells.gel.data.table.Table;
-import net.splitcells.gel.data.lookup.LookupComponents;
+import net.splitcells.gel.data.lookup.LookupMethods;
 import net.splitcells.gel.data.view.Line;
 import net.splitcells.gel.data.view.View;
 
-public interface ColumnView<T> extends ListView<T>, LookupComponents<T> {
+public interface ColumnView<T> extends ListView<T>, LookupMethods<T> {
     /**
      * TODO PERFORMANCE
      *